refactor table to use into groups
